use `Effects` as generic activity for cyber/space
